+++ CHANGELOG.md
@@ -9,7 +9,15 @@ Starting v0.13.0 for Microsoft.Azure.WebJobs.Extensions.OpenAI.Kusto, it will ma
 
 Starting v0.1.0 for Microsoft.Azure.WebJobs.Extensions.OpenAI.AzureAISearch, it will maintain its own [Changelog](./src/WebJobs.Extensions.OpenAI)
 
-## v0.14.0 - Unreleased
+## v0.15.0 - Unreleased
+
+### Changed
+
+- Added EmbeddingsStore binding and split SemanticSearch binding into two bindings: EmbeddingsStore and SemanticSearch
+- Added support for azure openai key authentication back (use of managed identity is always recommended)
+- Changed default embeddings model from `text-embedding-3-small` to `text-embedding-ada-002` as it is available in all regions of Azure.
+
+## v0.14.0 - 2024/04/24
 
 ### Changes